pyproject.toml
[project]
name = "yaade"
version = "0.1.0"
description = "Yaade (यादें) - Local MCP-compatible memory storage for AI agents with semantic search"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Reyansh Gupta"}
]
keywords = ["ai", "memory", "mcp", "claude", "vector-database", "embeddings"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi[standard]>=0.116.1",
    "mcp[cli]>=1.12.3",
    "sentence-transformers>=5.0.0",
    "pydantic-settings>=2.5.2",
    "mcp-server>=0.1.4",
    "numpy>=1.24.0",
    "textual>=0.70.0",
    "rich>=13.0.0",
]

[project.urls]
Homepage = "https://github.com/reyanshgupta/yaade"
Repository = "https://github.com/reyanshgupta/yaade"
Issues = "https://github.com/reyanshgupta/yaade/issues"
Documentation = "https://github.com/reyanshgupta/yaade#readme"

[project.scripts]
yaade = "app.cli:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
